Refactor test_publishers setUpTestData method and signals.py renamed to signal_handlers.py
